refactor : 장바구니, 상세화면 추가에 따른 패키지 구조변경
