+++ app/src/main/java/nextstep/shoppingcart/ProductListActivity.kt
@@ -3,12 +3,12 @@ package nextstep.shoppingcart
 import android.os.Bundle
 import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
+import nextstep.shoppingcart.model.product.ProductListUiState
+import nextstep.shoppingcart.ui.ProductListScreen
+import nextstep.shoppingcart.ui.getProductsTestData
 import nextstep.shoppingcart.ui.theme.ShoppingCartTheme
-import nextstep.shoppingcart.ui.theme.shopping.ProductListScreen
-import nextstep.shoppingcart.ui.theme.shopping.getProductsTestData
-import nextstep.shoppingcart.ui.theme.shopping.model.ProductListUiState
 
-class MainActivity : ComponentActivity() {
+class ProductListActivity : ComponentActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         val uiState = ProductListUiState.Products(getProductsTestData())
