+++ app/src/main/java/nextstep/shoppingcart/ui/component/ProductListContents.kt
@@ -1,4 +1,4 @@
-package nextstep.shoppingcart.ui.theme.shopping.component
+package nextstep.shoppingcart.ui.component
 
 import androidx.compose.foundation.lazy.grid.GridCells
 import androidx.compose.foundation.lazy.grid.LazyVerticalGrid
@@ -6,8 +6,8 @@ import androidx.compose.foundation.lazy.grid.items
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.tooling.preview.Preview
-import nextstep.shoppingcart.ui.theme.shopping.getProductsTestData
-import nextstep.shoppingcart.ui.theme.shopping.model.Product
+import nextstep.shoppingcart.model.product.Product
+import nextstep.shoppingcart.ui.getProductsTestData
 
 @Composable
 fun ProductListContents(productItems: List<Product>, modifier: Modifier = Modifier) {
